Limit scope of ProductDetails extension methods
